+++ vscode/src/completions/createVSCodeInlineCompletionItemProvider.ts
@@ -4,6 +4,7 @@ import { Configuration } from '@sourcegraph/cody-shared/src/configuration'
 import { FeatureFlagProvider } from '@sourcegraph/cody-shared/src/experimentation/FeatureFlagProvider'
 
 import { ContextProvider } from '../chat/ContextProvider'
+import { SectionObserver } from '../graph/section-observer'
 import { CodyStatusBar } from '../services/StatusBar'
 
 import { CodeCompletionsClient } from './client'
@@ -24,12 +25,15 @@ export async function createInlineCompletionItemProvider(
     const providerConfig = await createProviderConfig(config, client, featureFlagProvider)
     if (providerConfig) {
         const history = new VSCodeDocumentHistory()
+        const sectionObserver = config.autocompleteExperimentalGraphContext ? new SectionObserver() : undefined
+
         const completionsProvider = new InlineCompletionItemProvider({
             providerConfig,
             history,
             statusBar,
             getCodebaseContext: () => contextProvider.context,
             isEmbeddingsContextEnabled: config.autocompleteAdvancedEmbeddings,
+            graphContextFetcher: sectionObserver,
             completeSuggestWidgetSelection: config.autocompleteExperimentalCompleteSuggestWidgetSelection,
             featureFlagProvider,
         })
@@ -41,6 +45,9 @@ export async function createInlineCompletionItemProvider(
             vscode.languages.registerInlineCompletionItemProvider('*', completionsProvider),
             registerAutocompleteTraceView(completionsProvider)
         )
+        if (sectionObserver) {
+            disposables.push(sectionObserver)
+        }
     } else if (config.isRunningInsideAgent) {
         throw new Error(
             "Can't register completion provider because `providerConfig` evaluated to `null`. " +